zero: meta-csp: add image capture capability for scsat1
Implemented functionality to capture images (photos) using the camera on scsat1.

Modified the camera service to use `libcamerify` in front of the program to
ensure compatibility with libcamera.

Utilized `v4l2` for direct image capture from the camera.

This adds the capability to capture images from the camera on scsat1.

Signed-off-by: Gaetan Perrot <gaetan.perrot@spacecubics.com>